feat(refactor-mcts): Separate header file and add comments for doxygen
